Items Improvements
The Green and Red items play the fabled "walter white pan" and play a mini little dramatic animation when revealing which tiles can be used.

The blue item (jump) can now be cancelled, and costs an amount ROUGHLY equal to how far you jumped. So you might be able to farm it...but not by much.

We also significantly cleaned up the presentation of these items. I'd put a pin in this feature now. Well done.