Bugfix Suite 2
The levels number is no longer set to 0 every time on your 2nd play. Jesus, it took us like 3 months to get around to that...

The score changes like it's supposed to and that is reflected in Postgame stats.

Fixed a weird middle state between jumping and falling to death.

Updated the about/credits page.