Exit Button, Screen Resizing Confirmation
Exit game button in the main menu.

Screen resizing has a confirmation button now. One annoying thing is that it has to wait momentarily for the screen to actually resize before displaying the confirmation, and we just wait a constant number of seconds (no condition yet.) Hopefully it just works fine.